int64 remove

+++ backend/types/restaurant.go
@@ -1,8 +1,8 @@
 package types
 
 type Restaurant struct {
-	ID          int64  `json:"id"`
-	UserID      int64  `json:"userId"`
+	ID          int    `json:"id"`
+	UserID      int    `json:"userId"`
 	Name        string `json:"name"`
 	Location    string `json:"location"`
 	Description string `json:"description"`
@@ -13,20 +13,20 @@ type Restaurant struct {
 }
 
 type GetRestaurantReq struct {
-	UserID int64 `json:"userId" validate:"required" binding:"required"`
-	ID     int64 `json:"id" validate:"required" binding:"required"`
+	UserID int `json:"userId" validate:"required" binding:"required"`
+	ID     int `json:"id" validate:"required" binding:"required"`
 }
 
 type CreateRestaurantReq struct {
-	UserID      int64  `json:"userId" validate:"required" binding:"required"`
+	UserID      int    `json:"userId" validate:"required" binding:"required"`
 	Name        string `json:"name" validate:"required" binding:"required"`
 	Location    string `json:"location" validate:"required" binding:"required"`
 	Description string `json:"description" validate:"required" binding:"required"`
 }
 
 type UpdateRestaurantReq struct {
-	UserID      int64   `json:"userId"`
-	ID          int64   `json:"id"`
+	UserID      int     `json:"userId"`
+	ID          int     `json:"id"`
 	Name        *string `json:"name,omitempty"`
 	Location    *string `json:"location,omitempty"`
 	Description *string `json:"description,omitempty"`
@@ -35,6 +35,6 @@ type UpdateRestaurantReq struct {
 }
 
 type DeleteRestaurantReq struct {
-	UserID int64 `json:"userId" validate:"required" binding:"required"`
-	ID     int64 `json:"id" validate:"required" binding:"required"`
+	UserID int `json:"userId" validate:"required" binding:"required"`
+	ID     int `json:"id" validate:"required" binding:"required"`
 }
